+++ e_face_x4/app/connectors/__init__.py
@@ -1,6 +1,5 @@
 from .buspro import BusproConnector
 from .etherm import EThermConnector
+from .media import EkonexMediaConnector
 
-__all__ = ["BusproConnector", "EThermConnector"]
-
-__all__ = ["BusproConnector"]
+__all__ = ["BusproConnector", "EThermConnector", "EkonexMediaConnector"]